+++ src/main/java/ads/cip/services/GradeImpl.java
@@ -1,10 +1,12 @@
 package ads.cip.services;
 
+import ads.cip.exception.AlreadyExistsException;
+import ads.cip.exception.NotFoundException;
 import ads.cip.interfaces.IGrade;
 import ads.cip.model.GradeModel;
 import ads.cip.repository.GradeRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -12,7 +14,12 @@ import java.util.Optional;
 @Service
 public class GradeImpl implements IGrade {
 
-    private GradeRepository gradeRepository;
+    private final GradeRepository gradeRepository;
+    private static final String message = "Grade not found";
+
+    public GradeImpl(GradeRepository gradeRepository) {
+        this.gradeRepository = gradeRepository;
+    }
 
     @Override
     public List<GradeModel> getAllGrades() {
@@ -21,20 +28,32 @@ public class GradeImpl implements IGrade {
 
     @Override
     public Optional<GradeModel> getGradeById(String id) {
+        if(!gradeRepository.existsById(id)) {
+            throw new NotFoundException(message);
+        }
         return gradeRepository.findById(id);
     }
 
     @Override
+    @Transactional
     public GradeModel createGrade(GradeModel grade) {
-         return gradeRepository.save(grade);
+         if(gradeRepository.existsById(grade.getGradeId())){
+             throw new AlreadyExistsException("Grade already exists");
+         }
+        return gradeRepository.save(grade);
     }
 
     @Override
+    @Transactional
     public void deleteGrade(String id) {
+        if(!gradeRepository.existsById(id)) {
+            throw new NotFoundException(message);
+        }
         gradeRepository.deleteById(id);
     }
 
     @Override
+    @Transactional
     public GradeModel updateGrade(String id, GradeModel newGrade) {
         Optional<GradeModel> optionalGrade = gradeRepository.findById(id);
         if (optionalGrade.isPresent()) {
@@ -44,7 +63,8 @@ public class GradeImpl implements IGrade {
             existingGrade.setUserModel(newGrade.getUserModel());
             existingGrade.setExamModel(newGrade.getExamModel());
             return gradeRepository.save(existingGrade);
+        } else{
+            throw new NotFoundException(message);
         }
-        return null;
     }
 }
