+++ src/main/java/ads/cip/services/RoleImpl.java
@@ -1,10 +1,12 @@
 package ads.cip.services;
 
+import ads.cip.exception.AlreadyExistsException;
+import ads.cip.exception.NotFoundException;
 import ads.cip.interfaces.IRole;
 import ads.cip.model.RoleModel;
 import ads.cip.repository.RoleRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -12,7 +14,12 @@ import java.util.Optional;
 @Service
 public class RoleImpl implements IRole {
 
-    private RoleRepository roleRepository;
+    private final RoleRepository roleRepository;
+    private static final String message = "Role not found";
+
+    public RoleImpl(RoleRepository roleRepository) {
+        this.roleRepository = roleRepository;
+    }
 
     @Override
     public List<RoleModel> getAllRoles() {
@@ -21,20 +28,32 @@ public class RoleImpl implements IRole {
 
     @Override
     public Optional<RoleModel> getRoleById(String role) {
+        if(!roleRepository.existsById(role)){
+            throw new NotFoundException(message);
+        }
         return roleRepository.findById(role);
     }
 
     @Override
+    @Transactional
     public RoleModel saveRole(RoleModel role) {
+        if(roleRepository.existsById(role.getRoleId())){
+            throw new AlreadyExistsException("Role already exists");
+        }
         return roleRepository.save(role);
     }
 
     @Override
+    @Transactional
     public void deleteRole(String role) {
+        if(!roleRepository.existsById(role)){
+            throw new NotFoundException(message);
+        }
         roleRepository.deleteById(role);
     }
 
     @Override
+    @Transactional
     public RoleModel updateRole(String id, RoleModel newRole) {
         Optional<RoleModel> optionalRole = roleRepository.findById(id);
         if (optionalRole.isPresent()) {
@@ -42,7 +61,8 @@ public class RoleImpl implements IRole {
             existingRole.setRoleName(newRole.getRoleName());
             existingRole.setRoleDescription(newRole.getRoleDescription());
             return roleRepository.save(existingRole);
+        } else {
+            throw new NotFoundException(message);
         }
-        return null;
     }
 }
