+++ src/main/java/ads/cip/services/UserCertificationImpl.java
@@ -1,11 +1,13 @@
 package ads.cip.services;
 
+import ads.cip.exception.AlreadyExistsException;
+import ads.cip.exception.NotFoundException;
 import ads.cip.interfaces.IUserCertification;
 import ads.cip.model.UserCertificationId;
 import ads.cip.model.UserCertificationModel;
 import ads.cip.repository.UserCertificationRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -13,7 +15,12 @@ import java.util.Optional;
 @Service
 public class UserCertificationImpl implements IUserCertification {
 
-    private UserCertificationRepository userCertificationRepository;
+    private final UserCertificationRepository userCertificationRepository;
+    private static final String message = "User Certification not found";
+
+    public UserCertificationImpl(UserCertificationRepository userCertificationRepository) {
+        this.userCertificationRepository = userCertificationRepository;
+    }
 
     @Override
     public List<UserCertificationModel> getAllUserCertification() {
@@ -22,20 +29,32 @@ public class UserCertificationImpl implements IUserCertification {
 
     @Override
     public Optional<UserCertificationModel> getUserCertificationById(UserCertificationId id) {
+        if(!userCertificationRepository.existsById(id)){
+            throw new NotFoundException(message);
+        }
         return userCertificationRepository.findById(id);
     }
 
     @Override
-    public UserCertificationModel createUserCertification(UserCertificationModel userCertification) {
+    @Transactional
+    public UserCertificationModel createUserCertification(UserCertificationId id, UserCertificationModel userCertification) {
+        if(userCertificationRepository.existsById(id)){
+            throw new AlreadyExistsException("User Certification already exists");
+        }
         return userCertificationRepository.save(userCertification);
     }
 
     @Override
+    @Transactional
     public void deleteUserCertification(UserCertificationId  id) {
+        if(!userCertificationRepository.existsById(id)){
+            throw new NotFoundException(message);
+        }
         userCertificationRepository.deleteById(id);
     }
 
     @Override
+    @Transactional
     public UserCertificationModel updateUserCertification(UserCertificationId id, UserCertificationModel newUserCertification) {
         Optional<UserCertificationModel> existingUserCertificationOptional = userCertificationRepository.findById(id);
         if (existingUserCertificationOptional.isPresent()) {
@@ -47,7 +66,8 @@ public class UserCertificationImpl implements IUserCertification {
             existingUserCertification.setCertificationModel(newUserCertification.getCertificationModel());
 
             return userCertificationRepository.save(existingUserCertification);
+        } else {
+            throw new NotFoundException(message);
         }
-        return null;
     }
 }
