+++ src/main/java/ads/cip/services/UserImpl.java
@@ -1,9 +1,12 @@
 package ads.cip.services;
 
+import ads.cip.exception.AlreadyExistsException;
+import ads.cip.exception.NotFoundException;
 import ads.cip.interfaces.IUser;
 import ads.cip.model.UserModel;
 import ads.cip.repository.UserRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -12,7 +15,12 @@ import java.util.UUID;
 @Service
 public class UserImpl implements IUser {
 
-    private UserRepository userRepository;
+    private final UserRepository userRepository;
+    private static final String message = "User not found with id: ";
+
+    public UserImpl(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
 
     @Override
     public List<UserModel> getAllUsers() {
@@ -21,20 +29,32 @@ public class UserImpl implements IUser {
 
     @Override
     public Optional<UserModel> getUserById(UUID id) {
+        if(!userRepository.existsById(id)){
+            throw new NotFoundException(message + id);
+        }
         return userRepository.findById(id);
     }
 
     @Override
+    @Transactional
     public UserModel saveUser(UserModel user) {
+        if(userRepository.existsById(user.getUserId())){
+            throw new AlreadyExistsException("User already exists" + user.getUserId());
+        }
         return userRepository.save(user);
     }
 
     @Override
+    @Transactional
     public void deleteUser(UUID id) {
+        if(!userRepository.existsById(id)){
+            throw new NotFoundException(message + id);
+        }
         userRepository.deleteById(id);
     }
 
     @Override
+    @Transactional
     public UserModel updateUser(UUID id, UserModel newUser) {
         Optional<UserModel> optionalUser = userRepository.findById(id);
         if (optionalUser.isPresent()) {
@@ -52,7 +72,8 @@ public class UserImpl implements IUser {
             existingUser.setPosition(newUser.getPosition());
             existingUser.setRole(newUser.getRole());
             return userRepository.save(existingUser);
+        } else {
+            throw new NotFoundException(message + id);
         }
-        return null;
     }
 }
